merge

+++ ViajeFacil/app/Http/Controllers/GruposController.php
@@ -24,6 +24,17 @@ class GruposController extends Viajes
         $this->middleware('auth');  
     }
 
+    protected function grupoTienePostulacion($mi_grupo){
+        $postulacionesViajes = array();
+        $relacionDelGrupo = GruposViaje::where('id_grupo','=',$mi_grupo->id_grupo)->get();
+        $suma = 0;
+        foreach($relacionDelGrupo as $relacion)
+        {
+            $suma = Postulacion::where('id_viaje','=',$relacion->id_viaje)->where('estado_postulacion','=','aceptado')->count() + $suma;
+        }
+        return $suma;
+    }
+
     public function verDetallesGrupo($id_grupo)
     {
     	$user = Auth::user();
@@ -43,35 +54,60 @@ class GruposController extends Viajes
         ->with('viajes',$mis_viajes);
     }
 
+    public function modificarGrupo($id)
+    {
+        $viaje = Grupo::find($id);
+        $suma = $this->grupoTienePostulacion($viaje);
+        if($suma == 0){
+            $hora = explode(' ',$viaje->fecha)[1];
+            $vehiculos = parent::vehiculosUsuario();
+            return view('viajes.modificarViaje')->with('viaje',$viaje)
+            ->with('vehiculos',$vehiculos)
+            ->with('hora',$hora);
+        }else{
+            return redirect("/viajes/misViajes")->with('error', 'El viaje ya tiene postulaciones aceptadas y no se puede modificar');
+        }
+    }
+
     public function modificarGrupoId(Request $data)
     {
         $mi_grupo = Grupo::find($data['id_grupo']);
-        $mi_grupo->titulo = $data->input('titulo');
-        $mi_grupo->origen = $data->input('origen');
-        $mi_grupo->destino = $data->input('destino');
 
-        if (($mi_grupo->fecha != $data->fecha) || ($mi_grupo->tipo_viaje != $data->tipo_viaje)){
-            $grupos_viaje = GruposViaje::where('id_grupo', '=', $data['id_grupo']);
-            foreach ($grupos_viaje->get() as $dato)
-            {
-                parent::eliminarViaje($dato->id_viaje);
+        
+
+        $suma = $this->grupoTienePostulacion($mi_grupo);
+
+        if($suma == 0){
+            $mi_grupo->titulo = $data->input('titulo');
+            $mi_grupo->origen = $data->input('origen');
+            $mi_grupo->destino = $data->input('destino');
+
+            if (($mi_grupo->fecha != $data->fecha) || ($mi_grupo->tipo_viaje != $data->tipo_viaje)){
+                $grupos_viaje = GruposViaje::where('id_grupo', '=', $data['id_grupo']);
+                foreach ($grupos_viaje->get() as $dato)
+                {
+                    parent::eliminarViaje($dato->id_viaje);
+                }
+                $grupos_viaje->delete();
+                parent::createViajes($data);
             }
-            $grupos_viaje->delete();
-            parent::createViajes($data);
+
+            $date = explode('-',$data->fecha);
+            $carbonDate = Carbon::createFromDate($date[0],$date[1],$date[2]);
+            $carbonDate->setTimeFromTimeString($data->hora);
+            $date = date_create($carbonDate);
+            $mi_grupo->fecha = $date;
+
+            $mi_grupo->precio = $data->input('precio');
+            $mi_grupo->tipo_viaje = $data->input('tipo_viaje');
+
+            $mi_grupo->save();
+
+            return redirect("/viajes/misViajes")->with('mensaje', '¡El viaje ha sido modificado correctamente!');
+        }else{
+            return redirect("/viajes/misViajes")->with('error', 'El viaje ya tiene postulaciones aceptadas y no se puede modificar');
         }
 
-        $date = explode('-',$data->fecha);
-        $carbonDate = Carbon::createFromDate($date[0],$date[1],$date[2]);
-        $carbonDate->setTimeFromTimeString($data->hora);
-        $date = date_create($carbonDate);
-        $mi_grupo->fecha = $date;
-
-        $mi_grupo->precio = $data->input('precio');
-        $mi_grupo->tipo_viaje = $data->input('tipo_viaje');
-
-        $mi_grupo->save();
-
-        return redirect("/viajes/misViajes")->with('mensaje', '¡El viaje ha sido modificado correctamente!');
     }
 
     public function eliminarGrupo(Request $data)
